requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary>=2.9.5
django-extensions==3.2.1
python-dotenv==0.21.1